+++ NadekoBot/Modules/Music/MusicModule.cs
@@ -23,6 +23,7 @@ namespace NadekoBot.Modules.Music
 
         public MusicModule()
         {
+            commands.Add(new PlaylistCommands(this));
         }
 
         public override string Prefix { get; } = NadekoBot.Config.CommandPrefixes.Music;
@@ -566,152 +567,7 @@ namespace NadekoBot.Modules.Music
                         await e.Channel.SendMessage($"🎵🔁`Repeat playlist {(currentValue ? "enabled" : "disabled")}`").ConfigureAwait(false);
                     });
 
-                cgb.CreateCommand(Prefix + "save")
-                    .Description($"Saves a playlist under a certain name. Name must be no longer than 20 characters and mustn't contain dashes. | `{Prefix}save classical1`")
-                    .Parameter("name", ParameterType.Unparsed)
-                    .Do(async e =>
-                    {
-                        var name = e.GetArg("name")?.Trim();
-
-                        if (string.IsNullOrWhiteSpace(name) ||
-                            name.Length > 20 ||
-                            name.Contains("-"))
-                            return;
-
-                        MusicPlayer musicPlayer;
-                        if (!MusicPlayers.TryGetValue(e.Server, out musicPlayer))
-                            return;
-
-                        //to avoid concurrency issues
-                        var currentPlaylist = new List<Song>(musicPlayer.Playlist);
-                        var curSong = musicPlayer.CurrentSong;
-                        if (curSong != null)
-                            currentPlaylist.Insert(0, curSong);
-
-                        if (!currentPlaylist.Any())
-                            return;
-
-
-                        var songInfos = currentPlaylist.Select(s => new DataModels.SongInfo
-                        {
-                            Provider = s.SongInfo.Provider,
-                            ProviderType = (int)s.SongInfo.ProviderType,
-                            Title = s.SongInfo.Title,
-                            Uri = s.SongInfo.Uri,
-                            Query = s.SongInfo.Query,
-                        }).ToList();
-
-                        var playlist = new MusicPlaylist
-                        {
-                            CreatorId = (long)e.User.Id,
-                            CreatorName = e.User.Name,
-                            Name = name.ToLowerInvariant(),
-                        };
-                        DbHandler.Instance.SaveAll(songInfos);
-                        DbHandler.Instance.Save(playlist);
-                        DbHandler.Instance.InsertMany(songInfos.Select(s => new PlaylistSongInfo
-                        {
-                            PlaylistId = playlist.Id.Value,
-                            SongInfoId = s.Id.Value
-                        }));
-
-                        await e.Channel.SendMessage($"🎵 `Saved playlist as {name}-{playlist.Id}`").ConfigureAwait(false);
-
-                    });
-
-                cgb.CreateCommand(Prefix + "load")
-                    .Description($"Loads a playlist under a certain name.  | `{Prefix}load classical-1`")
-                    .Parameter("name", ParameterType.Unparsed)
-                    .Do(async e =>
-                    {
-                        var voiceCh = e.User.VoiceChannel;
-                        var textCh = e.Channel;
-                        if (voiceCh == null || voiceCh.Server != textCh.Server)
-                        {
-                            await textCh.SendMessage("💢 You need to be in a voice channel on this server.\n If you are already in a voice channel, try rejoining.").ConfigureAwait(false);
-                            return;
-                        }
-                        var name = e.GetArg("name")?.Trim().ToLowerInvariant();
-
-                        if (string.IsNullOrWhiteSpace(name))
-                            return;
-
-                        var parts = name.Split('-');
-                        if (parts.Length != 2)
-                            return;
-                        var playlistName = parts[0];
-
-                        int playlistNumber;
-                        if (!int.TryParse(parts[1], out playlistNumber))
-                            return;
-
-                        var playlist = DbHandler.Instance.FindOne<MusicPlaylist>(
-                            p => p.Id == playlistNumber);
-
-                        if (playlist == null)
-                        {
-                            await e.Channel.SendMessage("Can't find playlist under that name.").ConfigureAwait(false);
-                            return;
-                        }
-
-                        var psis = DbHandler.Instance.FindAll<PlaylistSongInfo>(psi =>
-                            psi.PlaylistId == playlist.Id);
-
-                        var songInfos = psis.Select(psi => DbHandler.Instance
-                            .FindOne<DataModels.SongInfo>(si => si.Id == psi.SongInfoId));
-
-                        await e.Channel.SendMessage($"`Attempting to load {songInfos.Count()} songs`").ConfigureAwait(false);
-                        foreach (var si in songInfos)
-                        {
-                            try
-                            {
-                                await QueueSong(e.User, textCh, voiceCh, si.Query, true, (MusicType)si.ProviderType).ConfigureAwait(false);
-                            }
-                            catch (PlaylistFullException)
-                            {
-                                break;
-                            }
-                            catch (Exception ex)
-                            {
-                                Console.WriteLine($"Failed QueueSong in load playlist. {ex}");
-                            }
-                        }
-                    });
-
-                cgb.CreateCommand(Prefix + "playlists")
-                    .Alias(Prefix + "pls")
-                    .Description($"Lists all playlists. Paginated. 20 per page. Default page is 0. |`{Prefix}pls 1`")
-                    .Parameter("num", ParameterType.Optional)
-                    .Do(e =>
-                    {
-                        int num = 0;
-                        int.TryParse(e.GetArg("num"), out num);
-                        if (num < 0)
-                            return;
-                        var result = DbHandler.Instance.GetPlaylistData(num);
-                        if (result.Count == 0)
-                            e.Channel.SendMessage($"`No saved playlists found on page {num}`").ConfigureAwait(false);
-                        else
-                            e.Channel.SendMessage($"```js\n--- List of saved playlists ---\n\n" + string.Join("\n", result.Select(r => $"'{r.Name}-{r.Id}' by {r.Creator} ({r.SongCnt} songs)")) + $"\n\n        --- Page {num} ---```").ConfigureAwait(false);
-                    });
-
-                cgb.CreateCommand(Prefix + "deleteplaylist")
-                    .Alias(Prefix + "delpls")
-                    .Description($"Deletes a saved playlist. Only if you made it or if you are the bot owner. | `{Prefix}delpls animu-5`")
-                    .Parameter("pl", ParameterType.Required)
-                    .Do(async e =>
-                    {
-                        var pl = e.GetArg("pl").Trim().Split('-')[1];
-                        if (string.IsNullOrWhiteSpace(pl))
-                            return;
-                        var plnum = int.Parse(pl);
-                        if (NadekoBot.IsOwner(e.User.Id))
-                            DbHandler.Instance.Delete<MusicPlaylist>(plnum);
-                        else
-                            DbHandler.Instance.DeleteWhere<MusicPlaylist>(mp => mp.Id == plnum && (long)e.User.Id == mp.CreatorId);
-                        await e.Channel.SendMessage("`Ok.` :ok:").ConfigureAwait(false);
-                    });
-
+                
                 cgb.CreateCommand(Prefix + "goto")
                     .Description("Goes to a specific time in seconds in a song.")
                     .Parameter("time")
